Updates the metro schedule
On some occasions a resource will be recreated during a simulated failure causing the migration to  miss this resource requiring a manual deployment. Expanding the schedule should reduce the likelihood of this. It is an issue in the migration schedule behaviour though and will hopefully be fixed in a future PX release.